+++ packages/server/src/resourceProviders/postgres/search/parameters/types.ts
@@ -1,0 +1,4 @@
+export type FilterSQLResult = {
+  query: string;
+  values: unknown[];
+};
